--- ViewModel/Product/Breadcrumbs.php
@@ -1,181 +1,0 @@
-<?php
-/**
- * Copyright © 2019 Studio Raz. All rights reserved.
- * For more information contact us at dev@studioraz.co.il
- * See COPYING_STUIDRAZ.txt for license details.
- */
-
-namespace SR\Breadcrumb\ViewModel\Product;
-
-use Magento\Catalog\Helper\Data;
-use Magento\Framework\App\Config\ScopeConfigInterface;
-use Magento\Framework\App\ObjectManager;
-use Magento\Framework\DataObject;
-use Magento\Framework\Serialize\Serializer\Json;
-use Magento\Framework\View\Element\Block\ArgumentInterface;
-use Magento\Framework\Escaper;
-
-/**
- * Product breadcrumbs view model.
- */
-class Breadcrumbs extends DataObject implements ArgumentInterface
-{
-    /**
-     * Catalog data.
-     *
-     * @var Data
-     */
-    private $catalogData;
-
-    /**
-     * @var ScopeConfigInterface
-     */
-    private $scopeConfig;
-
-    /**
-     * @var Escaper
-     */
-    private $escaper;
-
-    /**
-     * @param Data $catalogData
-     * @param ScopeConfigInterface $scopeConfig
-     * @param Json|null $json
-     * @param Escaper|null $escaper
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
-     */
-    public function __construct(
-        Data $catalogData,
-        ScopeConfigInterface $scopeConfig,
-        Json $json = null,
-        Escaper $escaper = null
-    ) {
-        parent::__construct();
-
-        $this->catalogData = $catalogData;
-        $this->scopeConfig = $scopeConfig;
-        $this->escaper = $escaper ?: ObjectManager::getInstance()->get(Escaper::class);
-    }
-
-    /**
-     * Returns category URL suffix.
-     *
-     * @return mixed
-     */
-    public function getCategoryUrlSuffix()
-    {
-        return $this->scopeConfig->getValue(
-            'catalog/seo/category_url_suffix',
-            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
-        );
-    }
-
-    /**
-     * Checks if categories path is used for product URLs.
-     *
-     * @return bool
-     */
-    public function isCategoryUsedInProductUrl(): bool
-    {
-        return $this->scopeConfig->isSetFlag(
-            'catalog/seo/product_use_categories',
-            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
-        );
-    }
-
-    /**
-     * Returns product name.
-     *
-     * @return string
-     */
-    public function getProductName(): string
-    {
-        return $this->catalogData->getProduct() !== null
-            ? $this->catalogData->getProduct()->getName()
-            : '';
-    }
-
-    /**
-     * Returns breadcrumb json with html escaped names
-     *
-     * @return string
-     */
-    public function getJsonConfigurationHtmlEscaped() : string
-    {
-        return json_encode(
-            [
-                'breadcrumbs' => [
-                    'categoryUrlSuffix' => $this->escaper->escapeHtml($this->getCategoryUrlSuffix()),
-                    'useCategoryPathInUrl' => (int)$this->isCategoryUsedInProductUrl(),
-                    'product' => $this->escaper->escapeHtml($this->getProductName()),
-                    'longestCategoryPath' =>$this->getLongestCategoryPath($this->catalogData->getProduct())
-                ]
-            ],
-            JSON_HEX_TAG
-        );
-    }
-
-    /**
-     * Returns breadcrumb json.
-     *
-     * @return string
-     * @deprecated 103.0.0 in favor of new method with name {suffix}Html{postfix}()
-     */
-    public function getJsonConfiguration()
-    {
-        return $this->getJsonConfigurationHtmlEscaped();
-    }
-
-    public function getLongestCategoryPath($product)
-    {
-        /** @var \Magento\Catalog\Model\ResourceModel\Category\Collection $collection */
-        $collection = $product->getCategoryCollection();
-        $collection
-            ->addAttributeToSelect('is_active')
-            ->addAttributeToSelect('name')
-            ->setOrder('level', 'DESC');
-        $pool           = [];
-        $targetCategory = null;
-        /** @var \Magento\Catalog\Model\Category $category */
-        foreach ($collection as $category) {
-            $pool[$category->getId()] = $category;
-            if (!$category->getIsActive()) {
-                continue;
-            }
-            // all parent categories must be active
-            $child = $category;
-            try {
-                while ($child->getLevel() > 1 && $parent = $child->getParentCategory()) {
-                    $pool[$parent->getId()] = $parent;
-                    if (!$parent->getIsActive()) {
-                        $category = null;
-                        break;
-                    }
-                    $child = $parent;
-                }
-            } catch (\Exception $e) {
-                // Not found exception is possible (corrupted data in DB)
-                $category = null;
-            }
-            if ($category) {
-                $targetCategory = $category;
-                break;
-            }
-        }
-        $path = [];
-        if ($targetCategory) {
-            $pathInStore = $category->getPathInStore();
-            $pathIds     = array_reverse(explode(',', $pathInStore));
-            foreach ($pathIds as $categoryId) {
-                if (isset($pool[$categoryId]) && $pool[$categoryId]->getName()) {
-                    $category = $pool[$categoryId];
-                    $path[] = [
-                        'label' => $category->getName(),
-                        'link'  => $category->getUrl(),
-                    ];
-                }
-            }
-        }
-        return $path;
-    }
-}
