+++ etc/module.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0"?>
 <!--
 /**
- * Copyright © 2019 Studio Raz. All rights reserved.
- * For more information contact us at dev@studioraz.co.il
- * See COPYING_STUIDRAZ.txt for license details.
+ * Copyright © 2020 Studio Raz. All rights reserved.
+ * See LICENSE.txt for license details.
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
