--- view/frontend/layout/catalog_product_view.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-
-<page layout="1column" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <!-- Replacing default Magento_Catalog viewmodel by ours -->
-        <referenceBlock name="breadcrumbs">
-            <arguments>
-                <argument name="viewModel" xsi:type="object">SR\Breadcrumb\ViewModel\Product\Breadcrumbs</argument>
-            </arguments>
-        </referenceBlock>
-    </body>
-</page>
