+++ view/frontend/requirejs-config.js
@@ -1,9 +1,14 @@
-var config = {
-    map: {
-        '*': {
-            'Magento_Catalog/js/product/breadcrumbs':
-                'SR_Breadcrumb/js/product/breadcrumbs',
-        },
+/**
+ * Copyright © 2020 Studio Raz. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
 
+var config = {
+    config: {
+        mixins: {
+            'Magento_Theme/js/view/breadcrumbs': {
+                'SR_Breadcrumb/js/product/breadcrumbs-mixin': true
+            }
+        }
     }
 };
